Amending$

+++ src/main/java/io/management/ua/utility/repository/UserDetailsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserDetailsRepository extends JpaRepository<UserDetailsModel, Integer> {
+public interface UserDetailsRepository extends JpaRepository<UserDetailsModel, Long> {
     Optional<UserDetailsModel> findAbstractUserModelByUsername(String username);
 }
